+++ core/renderer.ts
@@ -6,6 +6,7 @@ import { getBasename, getPageUrl } from "./utils/page_url.ts";
 import { getPageDate } from "./utils/page_date.ts";
 import { Page } from "./file.ts";
 import { posix } from "../deps/path.ts";
+import { log } from "./utils/log.ts";
 
 import type { Content, Data, RawData } from "./file.ts";
 import type Processors from "./processors.ts";
@@ -168,9 +169,11 @@ export default class Renderer {
               page.content = content;
             }
           } catch (cause) {
-            throw new Error(`Error rendering the page: ${page.sourcePath}`, {
-              cause,
-            });
+            log.error(
+              `[rendering] Error rendering the page ${page.sourcePath}: ${
+                (cause as Error).message
+              }`,
+            );
           }
         },
       );
@@ -191,9 +194,10 @@ export default class Renderer {
             page.data.children as Content,
           );
         } catch (cause) {
-          throw new Error(
-            `Error rendering the layout of the page ${page.sourcePath}`,
-            { cause },
+          log.error(
+            `[rendering] Error rendering the layout of the page ${page.sourcePath}: ${
+              (cause as Error).message
+            }`,
           );
         }
       },